refactor: Extract createResultElement process to function
